+++ src/main/java/com/rocs/infirmary/application/data/model/inventory/medicine/Medicine.java
@@ -8,7 +8,7 @@ import java.sql.Timestamp;
 
 public class Medicine extends Inventory{
 
-    private long medicineId;
+    private Long medicineId;
 
     private String itemName;
 
@@ -24,7 +24,7 @@ public class Medicine extends Inventory{
 
     }
 
-    public Medicine(int medicineId, String itemName, String description, Timestamp expirationDate,BooleanProperty isSelected) {
+    public Medicine(Long medicineId, String itemName, String description, Timestamp expirationDate, BooleanProperty isSelected) {
         this.medicineId = medicineId;
         this.itemName = itemName;
         this.description = description;
